src/test/java/com/example/demo/shop/products/PostgresProductsRepositoryTest.java
package com.example.demo.shop.products;

import com.example.demo.DemoApplication;
import com.example.demo.domain.Product;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.junit.MockitoJUnitRunner;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.Optional;

import static org.assertj.core.api.Assertions.assertThat;

@ActiveProfiles("test")
@RunWith(SpringRunner.class)
@SpringBootTest
@AutoConfigureTestDatabase(replace= AutoConfigureTestDatabase.Replace.NONE)
public class PostgresProductsRepositoryTest {

    @Autowired
    PostgresProductsRepository repository;

    @Test
    public void shouldGetOneProduct()
    {
        final Optional<Product> product = repository.find (1);

        assertThat(product).isNotNull ().isNotEmpty ();
    }

}